requirements.txt
databricks-sdk>=0.20.0
httpx>=0.25.0
openai>=1.3.0
python-dotenv>=1.0.0
